+++ .gitignore
@@ -1,4 +1,5 @@
 token.json
 credentials.json
 *.exe
+cmd/uzcalendar/uzcalendar
