+++ cmd/wasm/main.go
@@ -1,0 +1,1 @@
+package wasm
