+++ pkg/schedules/schedules_test.go
@@ -62,7 +62,7 @@ func TestParseSchedule(t *testing.T) {
 func TestMain(t *testing.M) {
 	// Run everything from project root folder
 	_, filename, _, _ := runtime.Caller(0)
-	dir := path.Join(path.Dir(filename), "..")
+	dir := path.Join(path.Dir(filename), "..", "..")
 	err := os.Chdir(dir)
 	if err != nil {
 		fmt.Println("error returning to root folder: ", err)
